refactor app.py more
